+++ CHANGELOG.md
@@ -8,6 +8,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.5.0] - 2022-04-04
+
 ### Added
 
 * Add `--verbose` flag to control logging level (#28).
@@ -79,7 +81,8 @@ Initial release.
 
 <!-- Links -->
 
-[Unreleased]: https://github.com/wojdatto/pyzet/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/wojdatto/pyzet/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/wojdatto/pyzet/releases/tag/v0.5.0
 [0.4.0]: https://github.com/wojdatto/pyzet/releases/tag/v0.4.0
 [0.3.0]: https://github.com/wojdatto/pyzet/releases/tag/v0.3.0
 [0.2.0]: https://github.com/wojdatto/pyzet/releases/tag/v0.2.0